requirements.txt
google-cloud-storage
google-cloud-speech
pydub
librosa
numpy